% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CusuM.R
\name{CusuM}
\alias{CusuM}
\title{Cumulative Sum chart}
\usage{
CusuM(x)
}
\arguments{
\item{x}{a row vector with data to be monitored in interests sequentially listed
from left to right. The input could also be a matrix; each column represents all
observed data of different individuals at a specific time sequence.}
}
\value{
CUSUM chart
}
\description{
Calculating and plotting Cumulative Sum chart
}
\details{
Assume that each observations follows N(0,1), k = .75, h = 2.5, with ARL0 = 200,
upward shifts only
}
\examples{
set.seed(10082)
a <- rnorm(20, mean = 0, sd = 1)
b <- rnorm(20, mean = 1, sd = 1)
c <- c(a,b)
as.matrix(c)
c <- t(c)
CusuM(c)
}
\author{
Yuhang Zhou
}
